EP-813. Renamed model_configuration_service file to model_configuration_builder.
