+++ tvb_epilepsy/service/pse/pse_service.py
@@ -3,7 +3,7 @@ import numpy as np
 from abc import abstractmethod, ABCMeta
 from tvb_epilepsy.base.constants.model_constants import K_DEF, YC_DEF, I_EXT1_DEF, A_DEF, B_DEF
 from tvb_epilepsy.base.utils.log_error_utils import initialize_logger, raise_value_error
-from tvb_epilepsy.service.model_configuration_service import ModelConfigurationBuilder
+from tvb_epilepsy.service.model_configuration_builder import ModelConfigurationBuilder
 
 
 class ABCPSEService(object):
