+++ tvb_epilepsy/service/pse_service.py
@@ -12,7 +12,7 @@ from tvb_epilepsy.base.model.disease_hypothesis import DiseaseHypothesis
 from tvb_epilepsy.base.model.model_configuration import ModelConfiguration
 from tvb_epilepsy.service.simulator.simulator import ABCSimulator
 from tvb_epilepsy.service.epileptor_model_factory import model_build_dict
-from tvb_epilepsy.service.model_configuration_service import ModelConfigurationBuilder
+from tvb_epilepsy.service.model_configuration_builder import ModelConfigurationBuilder
 from tvb_epilepsy.service.lsa_service import LSAService
 from tvb_epilepsy.service.simulator.simulator_tvb import SimulatorTVB
 from tvb_epilepsy.service.simulator.simulator_custom import custom_model_builder
