+++ tvb_epilepsy/tests/test_custom_h5_reader.py
@@ -8,7 +8,7 @@ from tvb_epilepsy.io.h5_reader import H5Reader
 from tvb_epilepsy.io.h5_writer import H5Writer
 from tvb_epilepsy.service.hypothesis_builder import HypothesisBuilder
 from tvb_epilepsy.service.lsa_service import LSAService
-from tvb_epilepsy.service.model_configuration_service import ModelConfigurationBuilder
+from tvb_epilepsy.service.model_configuration_builder import ModelConfigurationBuilder
 from tvb_epilepsy.tests.base import get_temporary_folder, remove_temporary_test_files
 
 head_dir = "head2"
